Task Management Backend with AJAX
Basic tasks are able to be created (ones without rollover and recurrence).
The backend is platform independent. JSON objects are returned upon request that are
parsed on the client side then used to build the UI. The same API should work with
mobile applications and anything else that can create an HTTP connection.

The JS front end parses the JSON and builds the HTML layout for all of the tasks.